revert back to using guid type on guid fields
when we moved from using System.Guid for guid fields to using string it created an issue where older versions of the CruiseDAL could not read the guid value, and would default to the default guid value. When those guids would get saved back it would cause unique constrain failures. So now we need to go back to using System.Guid