CHANGE separate out V2 updater out of CruiseDAL.Core
I realized that CruiseDAL.Core version id wasn't being updated with changes to V2 updater. To reduce possibility of this becoming an issue in the future I separated out the V2 updater into a separate project. Note: V2 Updater was kept separate from CruiseDAL.V2 so that the upconvertor wouldn't need a dependency on CruiseDAL.V2.